reports

+++ src/database/migrations/20211213105353-create-feed.js
@@ -13,10 +13,10 @@ module.exports = {
       },
       
       unit: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       price: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DECIMAL
       },
       measurement: {
         type: Sequelize.STRING
